third_party/blink/web_tests/dom/mutation-event-tests/dom/node/DOMNodeRemovedEvent-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMNodeRemoved' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
This test checks that DOMNodeRemovedEvent is emitted once (and only once).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS count is 1
PASS successfullyParsed is true

TEST COMPLETE
